Added testing and final touches
This commit adds testing of the model and helper methods, as well as
acceptance testing for the main requirements of the webapp.
The final version has text formatted to allow the user to have up to 2
favourite languages, and a list after that. It also takes care of
non-existent users, or users with no (public?) repos. It displays the
repo languages in a pie chart format, provided by the chartkick gem.

Additionally, it fixes a bug where the test suite was finishing
regardless of a fail in a framework.